requirements.txt
gTTS==2.5.4
moviepy==1.0.3
Pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
numpy<2
streamlit>=1.39.0
imageio-ffmpeg>=0.6.0
transformers>=4.30.0
torch>=2.0.0
torchaudio>=2.0.0
